Move valueModel static method to FieldModel
