+++ core/src/main/java/io/doov/core/dsl/field/types/FloatFieldInfo.java
@@ -15,7 +15,7 @@
  */
 package io.doov.core.dsl.field.types;
 
-import static io.doov.core.dsl.impl.DefaultFunction.valueModel;
+import static io.doov.core.FieldModel.valueModel;
 import static io.doov.core.dsl.meta.predicate.FieldMetadata.fieldMetadata;
 
 import io.doov.core.FieldInfo;
